NewAvatar/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.25
celery-with-redis==3.0
certifi==2017.11.5
chardet==3.0.4
construct==2.5.3
diff-match-patch==20121119
Django==1.11.7
django-celery==3.2.1
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.5.1
django-tinymce==2.6.0
et-xmlfile==1.0.1
future==0.16.0
haystack==0.42
httplib2==0.9.2
idna==2.6
image==1.5.16
jdcal==1.3
jieba==0.38
kombu==3.0.37
lxml==4.1.1
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
packaging==16.8
pefile==2017.9.3
Pillow==4.3.0
pymongo==3.4.0
PyMySQL==0.7.10
pyparsing==2.2.0
python-ptrace==0.9.3
pytz==2016.10
PyYAML==3.12
redis==2.10.5
requests==2.18.4
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
uWSGI==2.0.15
vine==1.1.3
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
